Add dynamic server list rendering in server component
The static rendering of a single server has been replaced with a dynamic list rendering in servers.component.html. This allows multiple servers to be displayed instead of just one. The corresponding changes have been made in servers.component.ts where an array of servers has been introduced and the server name is added to the list when a server is created.