Implemented dynamic display details feature in server component
Introduced a new button for displaying server details as well as dynamic CSS classes in servers.component.html. The details can now be toggled and the list of times the button was clicked is displayed. Added a new array and a function to handle these changes in servers.component.ts and added relevant CSS styles in servers.component.css for a better user experience.