Improve Recipe model structure and initialization
The Recipe model's structure was updated to include read-only properties allowing controlled object creation through a static createRecipe factory method. This change makes the way we create a Recipe instances more explicit and prevents unwanted modifications. Also, in order to match with this modification in the Recipe model, we initialized a sample Recipe instance in the recipe-list.component.