Update layout and enrich contents in recipe-list.component
The layout of recipe-list.component was revised to be more presentable and informative. Instead of only having a Recipe-item component, additional HTML blocks were added to provide a "New Recipe" button, a Recipe's name and description, and an image placeholder for better user experience and interface clarity. Meanwhile, a recipe-item component remains, possibly reserved for future enhancements.