Update header component for interaction
The previous links in the header component, 'View Recipes' and 'Shopping List' are updated to provide interactions. Added an event emitter in the header component, which emits the string 'recipes' or 'shopping' upon click. This allows users to toggle between the 'View Recipes' and 'Shopping List' view by clicking on the corresponding link. Also updated the AppComponent to subscribe to this event and render the required view according to the emitted string. This provides a better user experience by allowing users to switch views.