Add DropdownDirective to app components
The DropdownDirective was added to app.module.ts and applied to certain elements in recipe-detail.component.html and header.component.html. Aided in building up interactive features in the app including ability to manage individual recipes and a dropdown interface for users. Also, testing files for DropdownDirective was created to ensure functionality and robustness.