Update onRemoveIngredient function to delete ingredient
Updated onRemoveIngredient function in recipe-edit.component.ts to actually remove the ingredient from the array instead of just logging a message by an added index parameter. This parameter is now passed to the method when the 'X' button is clicked in the HTML of the recipe-edit component.

Also updated the test for the onRemoveIngredient function in recipe-edit.component.spec.ts to check if the ingredient is properly removed. In recipe-list.component.ts, I added a ngOnDestroy method plus a property (recipesChangedSub) to subscribe and unsubscribe the recipesChanged event to prevent memory leaks.