Add HttpClientModule and DataStorageService
Added HttpClientModule to the imports of the main module to enable Http communication with the backend. HttpClientTestingModule was also imported for the test environment. To manage data storage and fetching, DataStorageService was created.

Two methods, 'storeRecipes' and 'fetchRecipes' were implemented to store and retrieve data from firebase respectively. In addition, 'storeRecipes' and 'fetchRecipes' methods were added to HeaderComponent to respond to 'Save Data' and 'Fetch Data' clicks in the dropdown menu. Recipe fixtures moved to a separate file to mimic database state.