Refactor `toObject()` methods into transformer functions
Removed `toObject()` method from the `Recipe` and `Ingredient` models and implemented transformer functions instead to transform an instance of `Recipe` or `Ingredient` into a plain JavaScript object. This was done to avoid the tight coupling of transformation logic with domain models. The transformation logic is now more modular and can be used in different places without necessarily depending on the given class methods.

Additionally, the resolver service was implemented for fetching recipes before activating the `RecipeDetailComponent` and `RecipeEditComponent` routes to ensure that the required data is available before the component gets initialized. Also, error handling was added to `onFetchData()` method in `HeaderComponent`, and a check was added in `storeRecipes()` method of `DataStorageService` to prevent sending an empty array of recipes to the server.

The updated transformation logic is tested in `recipe-edit.component.spec.ts` file.