Add authentication features and initial setup.
In this commit, a User model was introduced to represent an authenticated user in the AuthService. AuthService allows users to sign in and sign out. The User model holds user information such as email and has a token that gets revoked after a certain time. Navigation links in the header such as "View Recipes" are now visible only for authenticated users, whereas the "login" link is available for unauthenticated users. A logout button has also been added to the header for authenticated users. Changes have been made in Auth.component.ts and Auth.service.ts to navigate the user to the "/recipes" path after a successful login.

This refactor provides user authentication functionality. Upon successful login, the web app securely stores user info to provide a seamless experience. This would prevent unauthorized access and make the application more secure.