Fix logout routing and add logout logic
Fixed the routing on logout and added logout function in auth.service.ts. The Router was imported in the header.component.ts and was used in the onLogout function to navigate to the '/auth' page after logging out. The auth.service.ts file was modified to add a logout function which sets the user Behavior subject value to null, indicating there are no users currently authenticated.

The commit aims to handle users logging out by effectively navigating to a dedicated logout page and modifying user's state. Now, upon successfully logging out, users are redirected to the '/auth' page.