Add Alert Component and update toast style
Introduce a reusable Alert Component to display error messages in a consistent way across the application. This component has been included in the app.module.ts file for global usage. Updates were made to the styling of the toasters in the styles.css file for better visibility and readability. The Alert Component uses specific class names for different type of alerts like success, warning etc. in the alert.component.css file for flexibility and re-usability. Spec file for Alert Component is created along with HTML template for the component. Alert Component is then utilized in the auth.component.html to show error messages during user authentication. This improves usability by providing more informative and styled feedback to users.